requirements.txt
streamlit
pyrebase4
pandas
fpdf